Make lint_pkgbuild opt-in via MAKEPKG_LINT_PKGBUILD=1
Cygwin bash is struggling with the thousands of commands that get invoked
when lint_pkgbuild is run. This especially is annoying for split packages
because the calls get multiplied by the package count.

For example when running --printsrcinfo:

msys/brotli (4 packages)
* 39.8 seconds with linting
* 7 seconds without

msys/tar (1 package)
* 10.2 seconds with linting
* 2 seconds without

This changes the default to not lint the PKGBUILD and adds a
MAKEPKG_LINT_PKGBUILD env var that can be used to enable linting.
We can for example enable this in CI.